requirements.txt
apturl==0.5.2
attrs==21.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==5.0.9
defer==1.0.6
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==4.8.1
ipython-genutils==0.2.0
jsonschema==3.2.0
jupyter-core==4.7.1
kaleido==0.2.1
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmlite==0.36.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.4.2
monotonic==1.5
nbformat==5.1.3
netifaces==0.10.4
numba==0.53.1
numexpr==2.7.3
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.7
pandas==1.3.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.3.1
platformdirs==2.2.0
plotly==5.1.0
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.0
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.4.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
retrying==1.3.3
scipy==1.7.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
systemd-python==234
tenacity==8.0.1
tqdm==4.62.0
traitlets==5.1.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.7.0
virtualenv-clone==0.5.6
wadllib==1.3.3
xkit==0.0.0
zipp==3.5.0
